msm_iommu_hw-v1: use u64 instead of int cause 43bits

+++ drivers/iommu/msm_iommu_hw-v1.h
@@ -74,8 +74,8 @@
 
 #define SET_FIELD(addr, mask, shift, v) \
 do { \
-	int t = readl_relaxed(addr); \
-	writel_relaxed((t & ~((mask) << (shift))) + (((v) & \
+	u64 t = readl_relaxed(addr); \
+	writel_relaxed((((u64)t) & ~((mask) << (shift))) + (((v) & \
 			(mask)) << (shift)), addr); \
 } while (0)
